DELEG: allow local host as the second client
Allow the client option to specify the local host so
a second client is not needed in order to run the recall
tests. In order to use main client as the second client
the mount options must be different in order to have two
different mounts. Since NFSv3 does not have delegations
then using NFSv3 in the client option works well just
to recall the delegations.

Any of the following client names are treated as local:
  --client ":nfsversion=3"
  --client "127.0.0.1:nfsversion=3"
  --client "localhost:nfsversion=3"