IB: gracefully handle a missing SEND fragment
The reassembly table for the given destination queue
pair is initialized in the SEND First packet. If this
packet is missing then the reassembly table will not
exist while processing the SEND Middle and Last packets.
In order to avoid raising an exception, setdefault and
pop with a default value of an empty reassembly
table is used instead.