ALLOC: add method verify_fallocate
This method is used to verify fallocate works as
expected. It used to allocate as well as to deallocate
disk space. It makes sure that the call succeeds when
expected or fails with the expected error.  Also, it
verifies that the size of the file is correct after
an allocate or deallocate and whether the call to
fallocate succeeds or fails.