ALLOC: fix alloc06 and dealloc06 tests
Added option --free-blocks with a default value of 64
to use as a cushion of this number of blocks when trying
to allocate most of the disk space in the device instead
of using only 4 blocks. Also, the size to allocate or
deallocate is increased to make sure there is enough
range for testing.
Renamed the main test path to testfile to make sure
there is no confusion and the debug info is correct.
Finally, added debug info displaying the available disk
space after each operation to make it easier to debug.