DELEG: modify how test groups are defined
Use a main list of tests where each element is a tuple
(testname, bitmask) where bitmask gives all the groups
the test is part of. This way is a lot easier to add
new groups and are automatically sorted the way it is
sorted in the main list.
Also, add test groups read_deleg and write_deleg to run
all read delegation tests and all write delegation tests
respectively.