TestUtil: add method run_func and display assertion
Run function with the given arguments and return the
results. All positional arguments are passed to the
function while the named arguments change the behavior
of the method. Object attribute "oserror" is set to the
OSError object if the function fails.
Also, display function call and arguments where all
string objects are truncated to avoid having a very
large debug print of a function like write().