RDMA: initial commit for NFS-over-RDMA tests
Verify correct functionality of NFS-over-RDMA.
Tests are divided into three groups: basic, read and
write. The basic tests deal mostly with verifying NFS
packets using the reply chunk and some other basic RDMA
functionality. The read tests deal with verifying NFS
read which in turn verify the RDMA write functionality.
Finally, the write tests deal with verifying NFS write
which in turn verify the RDMA read functionality. Also,
if the NFS read or write is small enough the client
could not use the RDMA write or read functionality, but
instead could use the receive buffer and transfer the
data using the RDMA SEND operations.

Tests verify the RPCoRDMA layer is sent when necessary
and that the RDMA chunk lists are sent with the correct
information which includes the number of chunks, number
of segments in each chunk and the correct information
for each segment. Tests verify each segment information
is correct and their corresponding RDMA read or write
information which includes correct handle, virtual
offset, DMA length and the XDR position for the case of
RDMA reads. In addition, the correct number of RDMA I/O
fragments is also verified and their corresponding
lengths and packet sequence numbers.