HOST: fix wait_cmd for sudo started commands
The issue with stopping commands started with sudo is
that it could create descendant processes and this is
why the regular mechanisms to stop the command doesn't
work. The get_pids() methods is added to get all
descendant process ids for a given process id and thus
use this list to kill each descendant process before
killing the main process.