PKTT: add option to disable dissecting RPC replies
Add option to Pktt constructor to enable or disable
dissecting RPC replies. When dealing with very large
packet traces dissecting RPC replies require a lot
of memory because the RPC packet calls are saved in
order to dissect the replies. In cases where only
the RPC calls are needed, not dissecting RPC replies
will save a lot of memory and also will speed up
processing as well.