RDMA: fix long request or position-zero read chunk
When running with krb5i, a long request or PZRC is used
to send the whole request. The read request could have
multiple segments, thus if there is a small segment an
RDMA_Read_Response_Only is used. Verify reassembly could
be done on RDMA_Read_Response_Only if the last segment
is a small segment.

Also, the message type should be RDMA_NOMSG on a long
request so look for any packet having a read chunk.