RDMA: add test assertions to verify file data
Use non-RDMA mount at setup and cleanup. Since files are
created with a non-RDMA mount at setup, on all NFS read
tests, files are examined for data corruption by comparing
the data read with the pattern used for creating the files.
On all NFS write tests, the files created are examined for
data corruption by reading these files using a non-RDMA
mount and comparing the data with the pattern used when
written.