RDMA: correctly handle when PSN wraps around
Make sure to find the correct RDMA read/write fragments
when the PSN wraps around. A method was added to return
the correct match string for the different cases: PSN in
middle of the range [m, n), at the lower limit [0, n),
at the upper limit [m, MAX] and finally when the PSN
wraps around [m, MAX] & [0, n).