REXEC: fix results processing
When working with large latency networks, some file
operations take longer than the original timeout but
increasing this timeout to re-sync the results queue
slows down all calls. A transaction id is added to
avoid this re-sync timeout and it is used to cache
results for truly asynchronous results processing.
The sync_timeout is replaced with a simple timeout
of 30 seconds to accommodate high latency networks
but at the same time it does return the result as
fast as it becomes available.