RDMA: log missing IB requests
Modify sort_ib_fragments() to add missing requests to
the list and labeled them as such. Add assertion when
a request is missing thus avoiding cascading failures
on incorrect virtual addresses.
Also, set opcode of missing write requests according
to their sizes. If request has a size which is less
than or equal to iosize then it must be only fragment.