RDMA: log missing IB SEND fragments
This is a special case since the first and last SEND
cannot be missing because there would be no reassembly.
A partially reassembled NFS call or reply is needed to
search for SEND fragments thus SEND only, first and last
are always available in the fragment list.
Also, on SEND there is no way to know the PSN range or
how many SEND fragments should be in the packet trace
thus calculate the range from the first and last
fragments in the list.