RDMAinfo: support multiple iWarp read requests
Created new class RDMArequest to hold the iWarp request
to map sink to source. A method get_offset() in this
class translates the offset from sink to source.
Added a new dictionary of iWarp requests where the key
is given by the sink stag mapping to a list of
RDMArequest objects. Also, iWarp read fragments are
now added using the translated offset.