DELEG: fix recall24 test failures
When holding a read delegation and an open for writing
is sent, latest Linux servers grant a read delegation
on this open for writing. Since the client returned
the delegation before the new open, the script was
expecting to recover the open on CB_RECALL.
Also, use correct stateid in subsequent READs, this
comes from the second open stateid or the lock stateid.