POSIX: fix access tests when running as root
When running as root and the file is created as root,
access is granted for either R or W even when the file
does not have R or W permission.

Asking for multiple access bits when running as root
and the file is created as root, expect access unless
asking for X and file does not have X permissions.