PKTT: replace deprecated module parser with ast
Modify method _convert_match() to convert a string match
expression into a valid match expression to be evaluated
by eval(). All items specified as valid packet layers
are replaced with a call to the correct wrapper function.

Older Python releases do not define ast.unparse().
Create function unparse() with limited functionality but
enough for the matching language it is needed for method
match(). Function runs twice as fast as ast.unparse(),
so always use it regardless if it is defined or not on
the ast module.

Do not escape backslashes again on static method escape().